support frequencies above 32 bit limit: ~ 4.29 GHz
* added API functions rtlsdr_set_center_freq64(), rtlsdr_set_harmonic_rx(),
  rtlsdr_get_center_freq64()
* amended rtl_tcp protocol: SET_FREQ_HI32 = 0x56
* print frequency in MHz in double

+++ src/convenience/rtl_convenience.h
@@ -44,7 +44,7 @@ int nearest_gain(rtlsdr_dev_t *dev, int target_gain);
  * \return 0 on success
  */
 
-int verbose_set_frequency(rtlsdr_dev_t *dev, uint32_t frequency);
+int verbose_set_frequency(rtlsdr_dev_t *dev, uint64_t frequency);
 
 /*!
  * Set device sample rate and report status on stderr
